Chooses a default GWT source level to match the current java runtime.
Now the GWT default source compatibility level will try to match the
system property java.specification.version.

A small refactor in SourceLevel and a fix to apicheck have been made to
allow specifing the java source compatibility level when processing apis
in order to avoid a compile error on an old version of emul.Enum.

Change-Id: I69dc0e9b1ac0ecf9a40ee2a08d8d555319d0af6f