Update patch set 2
Patch Set 2:

I just noticed that the ApiChecker fails if you are running with source compatibility 7 due to the fact that Enum.java in emul does not compile in java 7.

I was thinking of updating the old source jar gwt25userApi. It is a minor update (adds a cast) that does not affect the api in any way:

   public final int compareTo(E other) {
     // TODO: will a bridge method do the cast for us?
     // if (this.getDeclaringClass() != other.getDeclaringClass()) {
     // throw new ClassCastException();
     // }
-    return this.ordinal - other.ordinal;
+    return this.ordinal - ((Enum) other).ordinal;
   }

Any thoughts about it?

Patch-set: 2
Label: Verified=0