Update patch set 2
Patch Set 2:

Haven't looked at the code yet, but it seems like we should rev the protocol version number and modify the code to support both the old and new way. On the client, the version can be set differently for the IE 6/7 permutation, and hopefully the version not in use can be optimized out. On the server it should support both.

But before we get into that, what's the motivation for this change? Most users of GWT-RPC don't care about the details about the wire format. What improvements do they get by changing it?

Patch-set: 2
Label: Verified=0