Update patch set 4
Patch Set 4:

We should decide how we're going to delegate events, since delegation is the whole point of this change.

On the one hand, when a handler receives a focus event, I would expect the event source to be the actual widget that received focus, and not a container for it. On the other hand, it's weird to add to add an event handler to one widget and have it actually be placed on a different widget.

It seems like in the calling code, it would be clearer to  be explicit and say foo.getDelegate().addFocusListener(...).
Or similarly, if you're calling a method that takes a HasFocusable, you should pass in foo.getDelegate().

This suggests we might not need FocusComposite at all, and instead the subclass should make getDelegate() public, possibly using a forwarding method with a more descriptive name. What was the original use case?

Patch-set: 4