Update patch set 4
Patch Set 4:

Composites are an encapsulation mechanism. For a widget, being a composite is an implementation detail that it doesn't necessarily want to expose outside. That is the reason in many case why they are going into all this trouble in the first place. ValueListBox and SuggestBox from SDK are good examples for that. What you are suggesting both breaks encapsulation and also other things like UiHandler or using Composite with widgets that are implemented using Composites.

If a composite want to expose its details, they can easily do that by providing an accessor to the internal widget and people can add handlers to that and the source would be the internal widget as expected.

Patch-set: 4