Update patch set 4
Patch Set 4:

I have once written a widget that does this foo.getDelegate().addXyzHandler() style but it comes with the cost that you can not use @UiHandler with it. If you are used to UiBinder its pretty annoying to not have @UiHandler support for a widget so at the end I actually added foo.addXyzHandler() methods to the widget and forward them to the delegate and obviously keep getDelegate() public so you are able to match against the event source.

It works ok'ish once everyone gets used to this style but API wise it just doesn't feel that great. First, the "problem" with making it UiBinder/UiHandler compatible is that developers just don't read the JavaDoc of the forwarding method foo.addXyzHandler() which states that they should use foo.getDelegate() when matching event sources. Thats because they don't call the method on their own when using UiHandler. Secondly, once you have these forwarding methods and a public getDelegate() you end up with two ways of adding handlers.

So I doubt that making Composite.getDelegate() public is a good choice. You end up with widgets that are similar to SuggestBox and DateBox that give access to their internals (which may result in deprecated methods in the future like SuggestBox.getTextBox()).

Patch-set: 4
Label: Verified=0