Update patch set 4
Patch Set 4:

It doesn't work with UIBinder because you are changing the expected semantics of a widget. So even without UIBinder case, this is still how it suppose to work.

You need to think in term of consumption. If a composite has multiple focusable objects and wants to expose focus API, it has different options depending on the specific scenario.
In a SuggestBox case, it would use the input the field as the logical reference for Focus from outside point of view even though autocomplete popup might be focusable.

For another widget that has multiple inputs and it really wants to really provide a Focus API, it has two options. Either it needs to synthesize it in a way that it looks like a single focusable piece from outside or it will need implement additional APIs in addition to HasFocus. 

Being said that, I don't ever remember trying to implement Focus interface on a UIBinder composed view. Focus concept is more intended to be used with widgets in contrast to a complex view.

Patch-set: 4