Update patch set 4
Patch Set 4:

Fair enough - I don't understand the particular use cases. My point is that we should make sure they work before committing to an API.

It seems like for the UIBinder case we're talking about PanelComposite and in that case the delegate is straightforward - it's the place where children are added. (It will still break encapsulation somewhat because when the child asks for its parent it will be different than what you'd expect.) I wonder what web components do?

I don't know if we can achieve complete encapsulation given our existing API's. Probably the best to hope for would be to avoid being too surprising. 

About the current API: do we necessarily want the delegate to be passed to initWidget? Maybe there should be a separate setDelegate() method instead.

Patch-set: 4