Update patch set 4
Patch Set 4:

It might be possible to intercept the events and update the source after they're generated:

class ClickSourceFixer extends ClickHandler {
  ClickHandler wrapped;
  Object newSource;
  void onClick(Event e) {
    Object oldSource = e.getSource();
    e.setSource(newSource);
    try {
      wrapped.onClick(e);
    } finally {
      e.setSource(oldSource);
    }
  }  
}

When someone calls foo.addEventHandler, the composite could wrap the handler before passing it to the delegate.

But this isn't transparent because someone might keep a reference to the event. To fix that we'd have to copy the event. If the composite has its own set of event handlers then the event only has to be translated once. Perhaps we could have a ForwardingHandlerManager that does this.

Patch-set: 4