Update patch set 3
Patch Set 3:

(1 comment)

Note that SuggestBox's event forwarding was deprecated in favor of exposing the internal TextBox, so you should do suggestBox.getTextBox().addFocusHandler(handler) instead of suggestBox.addFocusHandler(handler). It breaks encapsulation but I believe there are cases where it's impractical and causes more harm than good.

PanelComposite is a good example too: panels shouldn't be composites as it breaks the add(child) vs. getParent() symmetry, and fixing it would require too much work.

In the end, I tend to think that composites shouldn't expose low-level events from their wrapped widgets, and shouldn't be containers.

For events, either you only use high-level events such as SelectionEvent for SuggestBox, ValueChangeEvent for ValueListBox or ValuePicker, or OpenEvent and CloseEvent for DisclosurePanel.

For panels, you'd rather extend Panel, or not make a Composite and instead simply implement IsWidget and HasWidgets.ForIsWidget.

Note that we have precedents of composites that are panel and violate the principle of least surprise for getParent(): DisclosurePanel, TabBar and TabPanel are such widgets. Our existing widgets that delegate events however are either broken or have deprecated/removed the delegation.

Conclusion: it's a real mess, do whatever you think is best, but I'd rather go with only genericized Composite and ResizeComposite, possibly introducing the 'delegate' concept, and without FocusComposite and PanelComposite.

Patch-set: 3