Update patch set 2
Patch Set 2: Code-Review+1

Hmm, so, UiBinder only supports event handlers that extend EventHandler, and the error message in this case points to GwtEvent (so we can say UiBinder is tailored for GwtEvent and not the more generic Event), but this change wants to make it legal for an addXxxHandler(SomeGwtEvent) method to return a bindery HandlerRegistration (i.e. mixing c.g.g.event and c.g.web.bindery.event types in the same method).

That sounds weird.

I wonder if it wouldn't be better to either:

* put an error message when we detect c.g.web.bindery.event.shared.HandlerRegistration saying "did you mean com.google.gwt.event.shared.HandlerRegistration"
* or fully support c.g.web.bindery.event: mandate "assignable to c.g.w.b.event.shared.Event", don't look for EventHandler (only "is interface"), and thus accept any HandlerRegistration.

Patch-set: 2
Label: Code-Review=+1