Update patch set 4
Patch Set 4:

Since we already have ResizeComposite, I'm not sure that's a good example of how this change is an improvement? It seems like we should have another example.

For custom panels, how well does extending Panel work? Suppose you want a fancy panel that's based on a UIBinder template and can be used in a UIBinder template, with a hole where the children should go and some extra widgets that are encapsulated and aren't exposed as children in the panel API. How hard is that to do? What's the best way to make that easy?

Patch-set: 4
Label: Verified=0