Update patch set 5
Patch Set 5:

re: "We cannot get rid of the mass by ignoring it - and not taking an action is  exactly that."

Well, not taking action does have the advantage that nobody needs to migrate any code. But I agree with fixing things; I just want to make sure that any change we make works well and isn't a partially-working thing that people need to learn. We should understand it well enough that we can easily explain to other people why they should use it and it won't be "it seemed like a good idea at the time but I don't fully understand it."

Patch-set: 5