Update patch set 1
Patch Set 1:

I thought little bit more on this.

As we must always call specific addXYZHandler methods expecting specific handler classes, I think it is impossible to avoid boilerplate without codegen.

The minimal boilerplate that I found is my previous proposal of updating the source of the event (we can put a helper in composite for source override). Anybody who wants to be good citizen can choose to do that.

If we do code generation, we can simplify it a lot and everybody can use it:

  interface FocusForwardingHandlers extends HasXHandlers, HasYHandlers, ForwardingHandler;

  FocusForwardingHandlers helper;
  
  protected void initWidget(Widget widget, T delegate) {
    super.initWidget(widget, delegate);
    helper = GWT.create(FocusForwardingHandlers.class);
    helper.init(this, delegate);
  };

  public HandlerRegistration addFocusHandler(FocusHandler handler) {
    return helper.addHandler(handler);
  }

but writing the codegen is also a lot of work. I'm not sure if it will worth for just fixing the source of event.

Patch-set: 1
Reviewer: Gerrit User 5110 <5110@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0