Update patch set 5
Patch Set 5: Code-Review-1

(1 comment)

I tend to agree with Jens here (now that we see that in isolation, without the FocusComposite and PanelComposite)

When you call initWidget(Widget,T), you already have a variable holding the delegate (because you have to put it as a child of the "root widget"), and changing that variable to a field to make it accessible elsewhere in the class is a no-brainer (and if you use UiBinder to construct your Composite's internals, you already have that field as a @UiField, so the concept of delegate at the Composite level only comes as a duplicate).

The only use case for getDelegate is when you have someone extending your Composite, so they can gain direct access to that "central widget"; but a) extending a Composite is probably an anti-pattern and b) exposing your widget to subclasses is a no-brainer too.

I really think what we want is to have getWidget() return some 'T extends Widget'.

Patch-set: 5
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=-1