Fix for issue 1525
Fix for 'fly-over' (mouse over, but not clicked) problem (issue 1525) with
CustomButton that have a hover image. By setting capture on a mouseover
the mouseout is always triggered so the widget is not left in a state
were the hover image stays active even after the mouse left. The
isCapturing boolean is not set because that would also trigger the click
handling code.

Change-Id: I2d49cae5ce0f8b8248aa15347914a5cd8da95287