Update patch set 3
Patch Set 3: -Code-Review

(1 comment)

That starts to look good.

As discussed previously, I think we should bump the protocol version, and conditioning the server output (whether to use the .concat hack or not) the request version should be easy (in user/src/com/google/gwt/user/server/rpc/impl/ServerSerializationStreamWriter.java, there's already something about long serialization).

We could also just punt to when we remove support for IE6/7. The, we can either bump the protocol version again (version 9 would be identical to 8 except for the absence of the .concat hack) or change clients to throw an IncompatibleServiceException (but then stale clients running in IE6/7 will choke on deserialization of big arrays; not really a problem right?)

Patch-set: 3
Label: Code-Review=0