Update patch set 3
Patch Set 3:

"The main reason I put up this patch is that currently IE9 is leaking memory quite badly for each RPC call you make. The root cause is that eval() on IE9 is designed to keep everything evaled in memory until the browsers session dies. See https://code.google.com/p/google-web-toolkit/issues/detail?id=57366 for more details and discussion."

Thanks, this is a good reason. Please add it (or something like it) to the patch description.

Patch-set: 3