Update patch set 5
Patch Set 5:

> When you call initWidget(Widget,T), you already have a variable holding the delegate (because you have to put it as a child of the "root widget"), and changing that variable to a field to make it accessible elsewhere in the class is a no-brainer (and if you use UiBinder to construct your Composite's internals, you already have that field as a @UiField, so the concept of delegate at the Composite level only comes as a duplicate).

For simple composites, they will use the single initWidget and they might use the getDelegate for access to base widget but this is not really the use case to go with getDelegate and this is NOT why I introduced it.

> The only use case for getDelegate is when you have someone extending your Composite, so they can gain direct access to that "central widget"; but a) extending a Composite is probably an anti-pattern and b) exposing your widget to subclasses is a no-brainer too.

I think you totally missed the point. This is the reason why I'm saying we are going in circles. The conversation became too long and the original intention for the change started to get lost. 

I started working on this is, because it is not to possible safely extend Composite to introduce more reusable classes (inside or outside of SDK) that does more delegations (ResizeComposite, FocusComposite, PanelComposite, IndexPanelComposite etc). In our internal code there are already multiple of them, duplicated in different projects and written poorly.
In the GWT SDK, we don't have those reusable classes so we are also duplicating the code in a poor way. ValueListBox received the second patch since I joined the team.

Patch-set: 5