Update patch set 5
Patch Set 5:

Hi, Goktug and I talked about this. Goktug is pretty sure that delegation is necessary to make FocusComposite work properly, but it's rather difficult to explain why, so the conversation is getting frustrating.

So let's put this patch aside for a while. It would be helpful for someone to download the FocusComposite patch and attempt to make it work without delegation in Composite. Then we can compare solutions and go from there.

Patch-set: 5
Reviewer: Gerrit User 5010 <5010@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0