Update patch set 4
Patch Set 4:

Patch Set 4 now bumps the default version number and changes the versioning so that the server now always responds with the same version as the browser has requested. If JSON.parse() is not available, the client will always request version 7.

I didn't do anything to the concat part on the server side, if it kicks in the JSONUtils.safeEval() will fail and the eval() will be used instead. It could be refactored further to not kick in at all if the version < 8 but I think that could be done in another changeset.

Patch-set: 4
Label: Verified=0