Update patch set 4
Patch Set 4:

(6 comments)

Looking at the history, I stumbled on another case besides the .concat hack where the protocol does not emit JSON but rather JSON-like JavaScript: the ClientSerializationStreamReader in DevMode uses Rhino to eval() the response, and Rhino has a hard limit on the length of a string literal, so the server emits a concatenation of string literals in this case. See
https://code.google.com/p/google-web-toolkit/source/detail?r=10888

That means we'll want to:

* only emit those concatenated strings in getVersion() < 8
* parse the response without Rhino (and as pure JSON) when it's in version 8.

About detecting version 8, I wonder if we shouldn't change the response format to include the version outside the JSON. I have no idea what impact it would have on the code though, so maybe it's a bad idea and extracting the version from the JSON like you did is the way to go.

But much more importantly: I'd wait for our decision whether we continue to support IE6/7 before going any farther. If we drop IE6/7 support, we can simplify the client *and* server code.

Also, currently the server-side code supports back to version 5 of the GWT-RPC protocol; but that support is actually limited to accepting requests from old clients, because responses are always sent with the last known version (according to the changes made in this patch). Is this really needed? should backwards-compatibility be kept at all? only for the last 2 versions? (7 and 8 here) Note that clients will choke just after parsing the response because getVersion() doesn't exactly match what they know how to deal with, so supporting older requests on the server-side means that the RPC is executed but the client can't decode the response. I don't know what's best here (vs. failing early), but we're not really concerned here, as we're only changing the response format, so we can continue to support old requests; I don't think it's worth supporting clients older than v7 though (cached/running clients while the server is being updated/redeployed)

FYI, ServerSerializationStreamWriter has some code that depends on getVersion() in writeLong, which is why I though the changes would be limited to that class. Digging further, that code was added in https://code.google.com/p/google-web-toolkit/source/detail?r=8274 and simply copied from AbstractSerializationStreamWriter (shared between client and server), so I think it was an oversight.

FYI, backwards-compat on the server-side for decoding requests was added at https://code.google.com/p/google-web-toolkit/source/detail?r=8269

Patch-set: 4
Label: Verified=0