Update patch set 4
Patch Set 4:

(4 comments)

To add to the versioning mess, there are Google projects that are using GWT-RPC to serialize data without using the "RPC" part (via hacks). For example, GWT-RPC encoded data might be stored client-side or embedded into the HTML page served up at initial page load. I think they're just calling the static methods and not worrying about versioning since we haven't changed the protocol in years. So this could cause unexpected breakage for them. On the other hand if they're using a serialization policy they're probably used to it :-)

It seems like we should do this gradually and start by supporting 7 and 8 for decoding, and make the decision in the caller to use version 8 when we know it's safe?

We will need to announce the change for people using trunk and include it in release notes for the next stable version.

Patch-set: 4