Update patch set 4
Patch Set 4:

(1 comment)

We need to take a decision re. supporting version 7 in parallel to version 8 and how to do it (for those using the RPC class without doing remote calls).

When we'll drop support for IE6/7 (and thus need to continue supporting version 7 of the protocol) is being discussed by the steering committee, so we should IMO stop working on this patch pending that decision; unless Google wants to get this fixed sooner in master with a transition period (but then you could also do that in your branch).

If there's a transition period where both protocol versions have to be supported in parallel, then we need to decide how best to do it. I'd like to propose moving this discussion to gwt-contrib (generally speaking, we should try to only discuss in Gerrit about the code itself, and move other discussions to gwt-contrib, or ask the SC to make a decision, depending on the kind of thing that needs to be decided)

Patch-set: 4