Update patch set 2
Patch Set 2:

The first alternative is HyperlinkImpl. But this would introduce spaghetti inheritance because of the existing HyperlinkImpl's. And it clutters the History stuff across multiple classes.

Other alternative: Add it only to HistoryImpl and let Hyperlink use HistoryImpl directly. But this is not optimal because then we have to duplicate the HistoryImpl.init() error handling and it feels hacky to me.

I think a new method to generate a link for a given history token is the best way in the end because generating URLs is a part of implementing History mechanisms.

This change makes the code of gwt-pushstate simpler. However I have already a workaround for it. So it's not critical and we can take time to think over it.

Patch-set: 2
Reviewer: Gerrit User 5145 <5145@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0