Update patch set 2
Patch Set 2:

Yes, correct the mouseout isn't triggered in some browsers when the widget has a hover image. I guess by using capture theoretically it could move the problem. However, when the user clicks on the button, capture is also used. So the problem would also be triggered if the users clicks and moves out of the button. The fact the methods set/release capture are put in the else clause is to have cleaner code. Functionally the else could be removed because in both cases if isCapture is true the methods are also called. Without the else it would mean if isCapture is true the methods set/release would be called twice. Therefor I think the use of capture is a acceptable solution.

Patch-set: 2
Reviewer: Gerrit User 5490 <5490@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0