Update patch set 2
Patch Set 2:

AFAICT, no, they won't be "in sync", as it's the purpose of the SelectionChangeHandler to update the current value (this.value) with the one being selected (using the mouse or keyboard) in the CellList. This is specifically where we want to fire a ValueChangeEvent.

Patch-set: 2
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0