Update patch set 2
Patch Set 2:

I tested it and it doesn't seem native buttons work like that. If you mouse click on a native button, hold it, and move the mouse outside the button other elements do get mouseover and mouseout events.

I also created a test (not added here) with a native preview handler. It does work, except when you move the mouse (quickly) outside the browser window, in which case it can happen that no native preview event is triggered until you move the mouse inside the window again. However, it's also about 15 lines (of code), which seems a lot of code for a workaround for a small but annoying bug.

Patch-set: 2