Unary plus in Jsni code was always discarded.
Unary plus in JavaScript is only noop when applied to a
numeric value otherwise it behaves as a conversion to double
operator. The compiler incorrectly assumed that is was a noop
in all cases.

The fix was already made in r5951 by cromwellian but was lost
in a subsequent rollback (r5954).

Fixes issue 6373 and issue 3942.

Change-Id: I4d260d7fdc08b942f44b0275fb1c842be1c807fd