Update patch set 3
Patch Set 3: Code-Review+1

It would be nice to have a GWTTestCase (to be run in web-mode) which tests against this (I guess by reading the value and finding it is already set), in case there is a regression. When we start mucking around with the code-gen for incremental JS output, it be nice to catch regressions.

Patch-set: 3
Reviewer: Gerrit User 5001 <5001@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+1