Update patch set 4
Patch Set 4:

The typeMarker$ field gets overwritten with null and those everything breaks as it is assumed to be pointing to nullmethod().

After fixing that the serialization error disappears but a bunch of test cases start failing.

@Stephen, do you want to take over again?

My fix was to replace the if in endVisit(JField) (code below) but I am not sure if there is any other special field that we should be taking care of.

      if (x.getLiteralInitializer() != null) {
        if (x.isFinal()) {
          // setup the constant value
          accept(x.getLiteralInitializer());
        } else {
          // setup the default value, see Issue 380
          accept(x.getLiteralInitializer().getType().getDefaultValue());
        }
      } else if (x == program.getIndexedField("Object.typeMarker")) {
        // Special fields whose initialization is done somewhere else.
        push(null);
      } else {
        // setup a default value
        accept(x.getType().getDefaultValue());
      }

Patch-set: 4
Label: Verified=0