Update patch set 3
Patch Set 3: Code-Review-1

As an idea, looks good, but remember that JSOs are a particularly bad place to add new arbitrary methods, especially mid-release (i.e. "consider 3.0"). Any downstream code that extends Element and has a method with the same signature is going to break.

As an example (admittedly, one that I am biased toward), GXT has a Element subclass class that supports this very same method signature, hasClassName(String). Adding final methods to non-final classes in general must be done carefully - since in a JSO either all methods must be final or the type itself must be final, all new methods should be carefully scrutinized.

My vote would be that this would be implemented in a different way to not require new methods, at least for 2.x - breaking changes with downstream code isn't generally expected in minor releases (though that hasn't stopped us in the past, see JClassType about 2.2.0). The 3.0 release might be a better target for the patch as-is.

Patch-set: 3
Label: Code-Review=-1