Update patch set 3
Patch Set 3:

Hi Colin. I understand your concerns.

However, I like to remind that we are only promising API compatibility in maintenance releases like (2.5.1 to 2.5.2). Waiting major releases for these kind of patches would mean that delays up to 6 months which is not very practical and acceptable pace - especially for Google's use. Being said that, we are also trying to avoid changes that would cause too much refactoring to do an upgrade. In this case, the fix is pretty straightforward for most end user usages. If this one is a real big trouble for you, then we can wait for another 4-5 months but I would not prefer that =)

For reducing the pain in the future; I don't expect this class to grow much but I wonder if it is possible for you to compose instead of extend this kind of classes? If you can't, feel free to propose other changes that you expect to see in this base class instead of adding it your own. For example, I'm also planning to add toggleClassName.

Patch-set: 3