Update patch set 1
Patch Set 1:

After thinking about this overnight, here's my take on this. (Keep in mind that I'm not an expert in this stuff.)

Most existing users probably don't care about subpixel values and changing the API would be hugely disruptive with little gain for them. So shouldn't deprecate any existing API's and using x|0 to round off the measurements is probably the right fix.

For those users who do want subpixel values, I think it makes sense to introduce a SubpixelElement class which is exactly like Element except that the appropriate methods return doubles. Since this is a JSO, I think we can freely cast from one to the other? (I haven't actually done much with JSO's, and it seems like there might be issues with dev mode?)

Since the cast is probably awkward, introducing an Element.toSubpixelElement() method (much like Goktug suggested) that does the proper incantation seems like a nice fix. If it turns out we need to do the same thing for some subclasses then we use the same naming convention. For example, we might have ImageElement.toSubpixelImageElement() and it's a bit verbose but there's no conflict.

It's a bit ugly that we'll have both API's around forever but I think that's the way it goes; backward compatibility is more important. If someone really wants to start with a clean slate then they can use Elemental (when it's done) or Dart.

Patch-set: 1