Update patch set 1
Patch Set 1: Code-Review-1

I've had that patch sitting in my Git repo for mor ethan a year (the copyright for HasPaths read 2011 before I updated it to push to review).

I rebased it on top of the latest master but haven't re-run the test, which is why I'm giving myself a -1.

Patch-set: 1
Label: Code-Review=-1