Update patch set 2
Patch Set 2:

Per John Tamplin's point, I agree it seems unlikely, but at the time I did reproduce his issue, so I think it is valid, even if very rare.

I assume it's likely the doubles adding up to something that really should be 9, like 8.99999, but |0 treats it as 8.

The broken pyjamas link sucks--looks like they renamed the project to pyjs and moved to Github. I can't find the original issue.

That said, I'm fine leaving this as-is until/if a bug is reported.

Patch-set: 2