Update patch set 2
Patch Set 2:

@Stephen - when you say you reproduced it, do you mean you reproduced the subpixel problem or the problem of |0 not having the intended effect?

If you are doing solely integer math with doubles, as would be the case with GWT-generated code, there can be no roundoff.  Yes, if you add floats together to get something that should be an integer, it generally won't be because of the imprecise nature of floating point.  I don't think it is a problem that 8.9999* winds up as 8.

Patch-set: 2
Reviewer: Gerrit User 5100 <5100@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0