Update patch set 2
Patch Set 2:

[4] is shorter than [1]. The only question is how common it is. I haven't worked with subpixel values myself, so I don't know whether we're talking about a rare corner case or something people will use all the time.

It seems like a pixel (even a virtual one) is a small enough unit that we'd rarely need to work with anything smaller. But then again I'm not a designer.

So, any ideas as to what we're going to use this API for? I hate to add a lot of clutter if it won't be used much.

Patch-set: 2