Update patch set 4
Patch Set 4:

On the server side, I don't think we need a max version? If the client sends version 8 it should understand it and respond with version 8. Perhaps we could add an HTTP header to make it more likely that we send back the right version even if the message itself can't be parsed?

We do need a default version because of those unfortunate static methods in the API. Perhaps that can be configurable. See RemoteServiceServlet.getCodeServerPort for one possible way. I'm a bit wary of adding a mutable global variable, though, versus fixing the callers.

On the client side, a configuration property adds flexibility and we can set it for each permutation. It might be simpler just to set the version to use, rather than a max version. Perhaps the IE6/7 permutation should stop the compile with an error if it's set to 8?

Patch-set: 4
Label: Verified=0