Update patch set 1
Patch Set 1: Code-Review+2

Daniel created this, because I complained to him for submitting my patch before I have the time to test the implications internally =)

Anyway, I tested it now and I think it doesn't worth spending our limited resources to fix the implications. Besides that, it is not a bad idea to revert it globally as I documented in the issue 3812.

Patch-set: 1
Label: Code-Review=+2