Update patch set 5
Patch Set 5:

Okay, so patch set 5 takes a much simpler approach and does not change the current behaviour in any way. It only adds a system property "gwt.rpc.version" which can be set to change the behaviour to use JSON.parse. This I think, is a good intermediate solution until we bump the version permanently to 8 and allows customers to work around the memory leak.

I did not change how devmode evaluates the response (still using Rhino) and I think it deserves another changeset entirely since it is not related to fixing this memory leak in any way. 

The special cases (array concatenation, string concatenation) are still sent as Javascript even in version 8. In that case the client will try to parse with JSON.parse and fallback to eval. The memory leak will still persist there but I think those corner cases are so rare that they can at least for now be omitted.

Patch-set: 5
Label: Verified=0